requirement.txt
torch==1.13.0+cu116
torchvision==0.14.0+cu116
torchaudio==0.13.0+cu116
numpy==1.23.5
mmcv_full==1.7.1
mmdet==2.28.2
urllib3==1.26.16
pyquaternion==0.9.9
nuscenes-devkit==1.1.10
yapf==0.33.0
tensorboard==2.14.0
motmetrics==1.1.3
pandas==1.1.5
